external: Add option to use a keyfile.
The specified password is then hashed and used to create a mapping over
the keyfile, at which point the mapped contents of the keyfile are fed
into the mappings for the specified device.

IMPORTANT: This change is -not- backwards-compatible with older versions
of this utility, because the key-file is straight binary and may contain
newlines, and because specifying a key-file to 'cryptsetup' causes the
utility to skip hashing (in the old version a "sha512" hash was used,
but in the new version this would be ignored even if it was specified);
keep in mind that this means that even if you -don't- use a keyfile, the
hash will be skipped and thus the encrypted device won't mount if it was
formatted using an older version.

Basically, the internal refactoring that I did causes compatibility
issues with older versions; I made the changes because maintaining
backwards-compatibility would have been insane given the extra work and
ugliness it would have added to code not currently being used for
production.